changes due to prop link and agent API change

+++ app/models/Assessment.scala
@@ -21,7 +21,7 @@ import play.api.libs.json.{Json, Reads}
 import serialization.JsonFormats._
 
 case class Assessment(
-                       linkId: Int,
+                       linkId: Long,
                        asstRef: Long,
                        listYear: String,
                        uarn:Long,
